backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
sqlalchemy==2.0.25
alembic==1.13.1
redis==5.0.1
httpx==0.25.2
docker==6.1.3
aiofiles==23.2.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0.post1
sentry-sdk==1.39.0


# LLM Providers
openai>=1.12.0
anthropic>=0.25.0
aiohttp>=3.9.0

# MCP
sse-starlette>=2.0.0

# YAML
pyyaml>=6.0.0

# Encryption
cryptography>=42.0.0